feat: Implement Project Variables Dialog with CRUD functionality
- Added ProjectVariablesDialog component for managing project variables.
- Introduced hooks for handling local state and pending changes for project variables.
- Created NameInput and DefaultValueDisplay components for variable editing.
- Integrated variable management into the Homebar component.
- Updated OverlayUI to include ProjectVariableDialog and related hooks.
- Refactored DialogOptions type to a separate file for better organization.
- Removed TopBar component and its related hooks, consolidating functionality into OverlayUI.
- Adjusted imports and exports across components to reflect the new structure.
- Updated routes to import VersionDialog from the new location in OverlayUI.
