+++ ui/src/features/Editor/components/OverlayUI/components/ActionBar/hooks.ts
@@ -1,14 +1,6 @@
 import { useState } from "react";
 
-export type DialogOptions =
-  | "deploy"
-  | "share"
-  | "version"
-  | "assets"
-  | "debugStop"
-  | "projectVariables"
-  | "collaboration"
-  | undefined;
+import { DialogOptions } from "../../types";
 
 export default () => {
   const [showDialog, setShowDialog] = useState<DialogOptions>(undefined);
