+++ ui/src/features/Editor/components/OverlayUI/components/Homebar/components/ProjectVariablesDialog/VariableEditDialog/components/NumberEditor.tsx
