+++ ui/src/features/Editor/components/OverlayUI/components/Homebar/components/index.ts
@@ -1,1 +1,5 @@
 export { default as WorkflowsDropdown } from "./WorkflowsDropdown";
+export { default as CollaborationActionBar } from "./CollaborationActionBar";
+export { default as HomeMenu } from "./HomeMenu";
+export { default as ProjectVariableDialog } from "./ProjectVariablesDialog";
+export { default as Breadcrumb } from "./Breadcrumb";
