+++ ui/src/features/Editor/components/OverlayUI/components/index.ts
@@ -6,3 +6,5 @@ export { default as DebugLogs } from "./DebugPanel/DebugLogs";
 export { default as DebugPreview } from "./DebugPanel/DebugPreview";
 export { default as DebugPanel } from "./DebugPanel";
 export { default as Homebar } from "./Homebar";
+export { default as ActionBar } from "./ActionBar";
+export { default as DebugActionBar } from "./DebugActionBar";
