+++ ui/src/features/Editor/components/OverlayUI/types.ts
@@ -1,0 +1,9 @@
+export type DialogOptions =
+  | "deploy"
+  | "share"
+  | "version"
+  | "assets"
+  | "debugStop"
+  | "projectVariables"
+  | "collaboration"
+  | undefined;
