--- ui/src/features/Editor/components/TopBar/components/index.ts
@@ -1,6 +1,0 @@
-export { default as ActionBar } from "./ActionBar";
-export { default as DebugActionBar } from "./DebugActionBar";
-export { default as Breadcrumb } from "./Breadcrumb";
-export { default as HomeMenu } from "./HomeMenu";
-export { default as ProjectVariableDialog } from "./ProjectVariablesDialog";
-export { default as CollaborationActionBar } from "./CollaborationActionBar";
