+++ ui/src/features/Editor/components/index.ts
@@ -1,5 +1,4 @@
 export { default as OverlayUI } from "./OverlayUI";
-export { default as TopBar } from "./TopBar";
 export { default as WorkflowTabs } from "./WorkflowTabs";
 export { default as ParamsDialog } from "./ParamsDialog";
 export { default as NodeDeletionDialog } from "./NodeDeletionDialog";
