+++ ui/src/routes/workspaces.$workspaceId_.projects_.$projectId.lazy.tsx
@@ -8,7 +8,7 @@ import BasicBoiler from "@flow/components/BasicBoiler";
 import ErrorPage from "@flow/components/errors/ErrorPage";
 import { ProjectCorruptionError } from "@flow/errors";
 import Editor from "@flow/features/Editor";
-import { VersionDialog } from "@flow/features/Editor/components/TopBar/components/ActionBar/components/Version/VersionDialog";
+import { VersionDialog } from "@flow/features/Editor/components/OverlayUI/components/ActionBar/components/Version/VersionDialog";
 import {
   ProjectIdWrapper,
   WorkspaceIdWrapper,
